requirements.txt
openai>=1.98.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pathlib>=1.0.1
flask>=2.3.3
flask-cors>=4.0.0
anthropic>=0.25.0
# redis>=4.5.0
psutil>=5.9.0
gunicorn>=23.0.0
typing
logging